+++ src/main/java/nl/sidnlabs/pcap/packet/Packet.java
@@ -87,7 +87,7 @@ public class Packet {
   protected TcpHandshake tcpHandshake;
   // time in millis between time packet was sent by server
   // and when ack for packet was received in case of TCP use( tsSec + tsmicros)
-  protected long tcpPacketRtt = -1;
+  protected int tcpPacketRtt = -1;
   private boolean tcpRetransmission;
 
   private TCPFlow flow;
